requirements.txt
streamlit
pandas
PyYAML
openpyxl